[AIP-21] Move contrib.hooks.aws_hook providers.amazon.aws.hooks.aws_hook
