+++ airflow/contrib/hooks/aws_dynamodb_hook.py
@@ -21,8 +21,8 @@
 """
 This module contains the AWS DynamoDB hook
 """
-from airflow.contrib.hooks.aws_hook import AwsHook
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsDynamoDBHook(AwsHook):
