+++ airflow/providers/amazon/aws/hooks/datasync.py
@@ -21,8 +21,8 @@ Interact with AWS DataSync, using the AWS ``boto3`` library.
 
 import time
 
-from airflow.contrib.hooks.aws_hook import AwsHook
 from airflow.exceptions import AirflowBadRequest, AirflowException, AirflowTaskTimeout
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AWSDataSyncHook(AwsHook):
