+++ airflow/providers/amazon/aws/hooks/glue_catalog.py
@@ -20,7 +20,7 @@
 """
 This module contains AWS Glue Catalog Hook
 """
-from airflow.contrib.hooks.aws_hook import AwsHook
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsGlueCatalogHook(AwsHook):
