+++ airflow/providers/amazon/aws/hooks/kinesis.py
@@ -19,7 +19,7 @@
 """
 This module contains AWS Firehose hook
 """
-from airflow.contrib.hooks.aws_hook import AwsHook
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsFirehoseHook(AwsHook):
