+++ airflow/providers/amazon/aws/hooks/lambda_function.py
@@ -20,7 +20,7 @@
 """
 This module contains AWS Lambda hook
 """
-from airflow.contrib.hooks.aws_hook import AwsHook
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsLambdaHook(AwsHook):
