+++ airflow/providers/amazon/aws/hooks/logs.py
@@ -22,7 +22,7 @@ This module contains a hook (AwsLogsHook) with some very basic
 functionality for interacting with AWS CloudWatch.
 """
 
-from airflow.contrib.hooks.aws_hook import AwsHook
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsLogsHook(AwsHook):
