+++ airflow/providers/amazon/aws/hooks/s3.py
@@ -29,8 +29,8 @@ from urllib.parse import urlparse
 
 from botocore.exceptions import ClientError
 
-from airflow.contrib.hooks.aws_hook import AwsHook
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 def provide_bucket_name(func):
