+++ airflow/providers/amazon/aws/hooks/sns.py
@@ -22,7 +22,7 @@ This module contains AWS SNS hook
 """
 import json
 
-from airflow.contrib.hooks.aws_hook import AwsHook
+from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
 
 class AwsSnsHook(AwsHook):
