+++ airflow/providers/postgres/hooks/postgres.py
@@ -154,7 +154,7 @@ class PostgresHook(DbApiHook):
         or Redshift. Port is required. If none is provided, default is used for
         each service
         """
-        from airflow.contrib.hooks.aws_hook import AwsHook
+        from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
 
         redshift = conn.extra_dejson.get('redshift', False)
         aws_conn_id = conn.extra_dejson.get('aws_conn_id', 'aws_default')
