+++ tests/providers/mysql/hooks/test_mysql.py
@@ -153,7 +153,7 @@ class TestMySqlHookConn(unittest.TestCase):
         self.assertEqual(kwargs['ssl'], SSL_DICT)
 
     @mock.patch('airflow.providers.mysql.hooks.mysql.MySQLdb.connect')
-    @mock.patch('airflow.contrib.hooks.aws_hook.AwsHook.get_client_type')
+    @mock.patch('airflow.providers.amazon.aws.hooks.aws_hook.AwsHook.get_client_type')
     def test_get_conn_rds_iam(self, mock_client, mock_connect):
         self.connection.extra = '{"iam":true}'
         mock_client.return_value.generate_db_auth_token.return_value = 'aws_token'
